Lift the user selected movie id back to navigation root

+++ app/src/main/java/com/zg/netflixloginscreenjetpackcompose/ui/resusable_composables/HorizontalContinueWatchingMovieList.kt
@@ -9,13 +9,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.zg.netflixloginscreenjetpackcompose.data.models.Movie
 import com.zg.netflixloginscreenjetpackcompose.ui.list_items.ContinueWatchingListItem
-import com.zg.netflixloginscreenjetpackcompose.ui.list_items.MovieListItem
 import com.zg.netflixloginscreenjetpackcompose.ui.theme.MARGIN_MEDIUM
 import com.zg.netflixloginscreenjetpackcompose.ui.theme.MARGIN_MEDIUM_2
 import com.zg.netflixloginscreenjetpackcompose.ui.theme.NetflixCloneJetpackComposeTheme
 
 @Composable
-fun HorizontalContinueWatchingMovieList(movieList: List<Movie>, onTapMovie: () -> Unit, modifier: Modifier = Modifier) {
+fun HorizontalContinueWatchingMovieList(movieList: List<Movie>, onTapMovie: (Int) -> Unit, modifier: Modifier = Modifier) {
     LazyRow(
         horizontalArrangement = Arrangement.spacedBy(MARGIN_MEDIUM),
         contentPadding = PaddingValues(horizontal = MARGIN_MEDIUM_2),
